+++ js/util.js
@@ -1,3 +1,5 @@
+import {messages,names,descriptions,idCommentCount,creatPhotoCount} from './data.js';
+
 function getRandomInteger (min, max) {
   const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
   const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
@@ -15,4 +17,27 @@ const createIdGenerator = () => {
   };
 };
 
-export {getRandomInteger,getRandomArrayElement,createIdGenerator};
+const generateId = createIdGenerator();
+const generatePhotoId = createIdGenerator();
+
+const createComment = () => {
+  const comments = {
+    id: generateId(),
+    avatar: `img/avatar-${getRandomInteger(1,idCommentCount)}.svg`,
+    message: getRandomArrayElement(messages) + getRandomArrayElement(messages) ,
+    name: getRandomArrayElement(names),
+  };
+  return comments;
+};
+
+const creatPhoto = () => ({
+  id: generateId(),
+  url:`photos${generatePhotoId()}.jpg `,
+  description: getRandomArrayElement(descriptions),
+  like: getRandomInteger(15,200),
+  comment:createComment(),
+});
+
+const creatPhotos = Array.from({length:creatPhotoCount}, creatPhoto);
+
+export {creatPhotos};
